+++ packages/gamut/src/index.tsx
@@ -25,8 +25,8 @@ export * from './Form';
 export * from './GridForm';
 export * from './HiddenText';
 export * from './Typography/Text';
-export * from './Typography/DeprecatedHeading';
-export * from './Typography/DeprecatedText';
+export * from './Typography/HeadingDeprecated';
+export * from './Typography/TextDeprecated';
 export * from './HighlightedText';
 export * from './InputStepper';
 export * from './Interstitial';
