+++ packages/styleguide/stories/Atoms/Markdown/index.stories.mdx
@@ -1,4 +1,4 @@
-import { DeprecatedText, Markdown } from '@codecademy/gamut/src';
+import { TextDeprecated, Markdown } from '@codecademy/gamut/src';
 import title from '@codecademy/macros/lib/title.macro';
 import { Canvas, Meta, Props, Story } from '@storybook/addon-docs/dist/blocks';
 
@@ -68,7 +68,7 @@ If you need to override a link, iframe, or table, you also need to provide the `
 
 export const overrides = {
   a: {
-    component: DeprecatedText,
+    component: TextDeprecated,
   },
 };
