+++ packages/styleguide/stories/Molecules/SplitInterstitial/index.stories.mdx
@@ -1,5 +1,5 @@
 import {
-  DeprecatedHeading,
+  HeadingDeprecated,
   SplitInterstitial,
   Text,
 } from '@codecademy/gamut/src';
@@ -29,7 +29,7 @@ form on another.`,
         left={{
           children: (
             <>
-              <DeprecatedHeading as="h1">Important stuff!</DeprecatedHeading>
+              <HeadingDeprecated as="h1">Important stuff!</HeadingDeprecated>
               <Text>Some sort of information here, perhaps?</Text>
             </>
           ),
@@ -37,7 +37,7 @@ form on another.`,
         right={{
           children: (
             <>
-              <DeprecatedHeading as="h2">Less important...</DeprecatedHeading>
+              <HeadingDeprecated as="h2">Less important...</HeadingDeprecated>
               <Text>Some more information here, maybe?</Text>
             </>
           ),
@@ -60,7 +60,7 @@ You can override the built-in images with your own on a per-design basis.
         left={{
           children: (
             <>
-              <DeprecatedHeading as="h1">Important stuff!</DeprecatedHeading>
+              <HeadingDeprecated as="h1">Important stuff!</HeadingDeprecated>
               <Text>Some sort of information here, perhaps?</Text>
             </>
           ),
@@ -68,7 +68,7 @@ You can override the built-in images with your own on a per-design basis.
         right={{
           children: (
             <>
-              <DeprecatedHeading as="h2">Less important...</DeprecatedHeading>
+              <HeadingDeprecated as="h2">Less important...</HeadingDeprecated>
               <Text>Some more information here, maybe?</Text>
             </>
           ),
