+++ packages/styleguide/stories/Molecules/Tabs/index.stories.mdx
@@ -1,5 +1,5 @@
 import {
-  DeprecatedHeading,
+  HeadingDeprecated,
   Tab,
   TabList,
   TabPanel,
@@ -37,15 +37,15 @@ to show, but only one should be visible at a time.`,
           <Tab>Tab 3</Tab>
         </TabList>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 1</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 1</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 1. Yippee!</Text>
         </TabPanel>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 2</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 2</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 2. Yippee!</Text>
         </TabPanel>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 3</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 3</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 3. Yippee!</Text>
         </TabPanel>
       </Tabs>
@@ -75,15 +75,15 @@ instead notify of a request to change using `onChange`.
           <Tab>Tab 3</Tab>
         </TabList>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 2</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 2</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 2. Yippee!</Text>
         </TabPanel>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 1</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 1</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 1. Yippee!</Text>
         </TabPanel>
         <TabPanel>
-          <DeprecatedHeading as="h2">Welcome to Tab 3</DeprecatedHeading>
+          <HeadingDeprecated as="h2">Welcome to Tab 3</HeadingDeprecated>
           <Text>Hi there! I&apos;m the contents inside Tab 3. Yippee!</Text>
         </TabPanel>
       </Tabs>
